requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
yfinance>=0.2.18
google-generativeai>=0.3.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
nltk>=3.8.1
schedule>=1.2.0
aiohttp>=3.8.0
psutil>=5.9.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
scrapy>=2.11.0
twisted>=22.10.0
psycopg2-binary>=2.9.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0
flask-bcrypt>=1.0.0
flask-wtf>=1.1.0
wtforms>=3.0.0
email-validator>=2.0.0
